[expo-web-browser] Move TS code to expo-web-browser
