--- packages/expo-web-browser/src/ExpoModuleTemplateView.tsx
@@ -1,13 +1,0 @@
-import * as React from 'react';
-
-import { requireNativeViewManager } from 'expo-core';
-
-export default class ExpoWebBrowserView extends React.Component {
-  static NativeView = requireNativeViewManager('ExpoWebBrowserView');
-
-  render() {
-    return (
-      <ExpoWebBrowserView.NativeView />
-    );
-  }
-}
