+++ packages/expo-web-browser/src/ExpoWebBrowser.ts
@@ -1,0 +1,2 @@
+import { NativeModulesProxy } from 'expo-core';
+export default NativeModulesProxy.ExpoWebBrowser || {};
