+++ packages/expo-web-browser/src/ExpoWebBrowser.web.ts
@@ -1,5 +1,5 @@
 export default {
   get name() {
-    return 'ExponentWebBrowser';
+    return 'ExpoWebBrowser';
   },
 };
