--- packages/expo-web-browser/src/index.ts
@@ -1,9 +1,0 @@
-import { NativeModulesProxy } from 'expo-core';
-
-const { ExpoWebBrowser } = NativeModulesProxy;
-
-export { default as ExpoWebBrowserView } from './ExpoWebBrowserView';
-
-export async function someGreatMethodAsync(options: any) {
-  return await ExpoWebBrowser.someGreatMethodAsync(options);
-}
